Rearrange specs
